+++ src/busses/shunts.jl
@@ -1,21 +1,4 @@
 """
-    struct ShuntImpedance <: AbstractBus
-
-Required fields:
-- `bus::String`
-- `r::Real`  resistance in Ω
-- `x::Real`  reactance in Ω
-"""
-@with_kw struct ShuntImpedance <: AbstractBus
-    # required values
-    bus::String
-    r::Real
-    x::Real
-    # TODO matrices for MultiPhase models ?
-end
-
-
-"""
     struct ShuntAdmittance <: AbstractBus
 
 Required fields: